Increase the max filesize for media Currently we top out at 2.14 GB.
The following SQL can be used to convert:
begin;
update mediaencodingbinding set filesizebig = filesize;
alter table mediaencodingbinding rename filesize to filesize_old;
alter table mediaencodingbinding rename filesizebig to filesize;
alter table mediaencodingbinding drop filesize;
commit;

+++ sql/pgsql/tables/MediaEncodingBinding.sql
@@ -3,7 +3,7 @@ create table MediaEncodingBinding (
 	media_encoding integer not null references MediaEncoding(id),
 	media_type     integer not null references MediaType(id),
 
-	filesize integer,
+	filesize bigint,
 
 	on_cdn boolean not null default false,
